fix all typescript warnings (#10428)
* fix all typescript warnings

* clean up types

---------
