+++ services/QuillLMS/client/app/bundles/Diagnostic/components/questions/questions.jsx
@@ -8,7 +8,7 @@ import SelectSearch from 'react-select-search';
 import { fuzzySearch } from 'react-select-search';
 import { push } from 'react-router-redux';
 import respWithStatus from '../../libs/responseTools.js';
-import { submitResponseEdit, setUpdatedResponse, deleteResponse } from '../../actions/responses';
+import { submitResponseEdit, deleteResponse } from '../../actions/responses';
 import {
   Modal,
   ArchivedButton,
@@ -127,10 +127,6 @@ class Questions extends React.Component {
           questionUID: response.questionUID,
           gradeIndex: `unmatched${response.questionUID}`,
         };
-        sleep(150);
-        this.props.dispatch(
-          setUpdatedResponse(response.key, newValues)
-        );
       } else if (newMatchedResponse.response.parentID === undefined) {
         this.props.dispatch(
           deleteResponse(question.key, response.key)
