+++ services/QuillLMS/client/app/bundles/Grammar/components/questions/questions.tsx
@@ -2,11 +2,9 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import * as actions from '../../actions/questions';
 import _ from 'underscore';
-import { Link } from 'react-router';
 import QuestionListByConcept from './questionListByConcept'
 import checkAnswer from '../../libs/checkAnswer';
 import { push } from 'react-router-redux';
-import respWithStatus from '../../libs/responseTools.ts';
 import { submitResponseEdit, deleteResponse } from '../../actions/responses';
 import {
   Modal,
@@ -50,14 +48,12 @@ class Questions extends React.Component {
   }
 
   createNew() {
-    this.props.dispatch(actions.toggleNewQuestionModal());
   }
 
   submitNewQuestion() {
     const newQuestion = { name: this.refs.newQuestionName.value, };
     this.props.dispatch(actions.submitNewQuestion(newQuestion));
     this.refs.newQuestionName.value = '';
-    // this.props.dispatch(actions.toggleNewQuestionModal())
   }
 
   updateRematchedResponse(rid, vals) {
