+++ services/QuillLMS/client/app/bundles/Shared/components/questionList/linkListItem.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import {Link} from 'react-router'
+import { Link } from 'react-router-dom'
 
 // interface LinkListProps {
 //   basePath: string;
