+++ services/QuillLMS/client/app/bundles/Shared/components/renderForQuestions/renderThankYou.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 
 const ThankYou = (props: any) => (
   <section className="section">
