+++ services/QuillLMS/client/app/bundles/Shared/components/scoreAnalysis/questionRow.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 
 const QuestionRow = (props: any) => (
   <tr>
