--- services/QuillLMS/client/app/bundles/Shared/interfaces/misc.ts
@@ -1,13 +1,0 @@
-export interface UserIdsForEvent {
-  studentId: number,
-  teacherId: number
-}
-
-export interface NumberFilterInputProps {
-  handleChange: Function;
-  label: string;
-  column: {
-    filterValue: string;
-    id: string;
-  }
-}
